Stream index now refers to last sample + 1, as Guido intended.
